server/requirements.txt
# Specify PyTorch CPU index to ensure CPU-only version is prioritized
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

# API dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Authentication dependencies
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0

# Machine learning dependencies
torch==2.0.1
onnxruntime==1.17.1 # For ONNX model inference
opencv-python-headless==4.8.1.78
torchvision==0.15.2
# Image processing
pillow==10.0.1
numpy==1.24.3

# UI dependencies
gradio==4.29.0
gradio_imageslider 